requirements.txt
telethon==1.5.5
cryptg
incremental
twisted
python-dotenv
PySocks
psycopg2-binary
sqlalchemy
progressbar2
requests
asyncssh
aiohttp
telethon-session-sqlalchemy
